+++ lib/date/is.ts
@@ -4,9 +4,7 @@
  * @param {unknown} val - Value to verify
  */
 function isDate (val:unknown):val is Date {
-    /* eslint-disable-next-line */
-    /* @ts-ignore */
-    return val instanceof Date && !isNaN(val as number);
+    return val instanceof Date && val.getTime() === val.getTime();
 }
 
 export {isDate, isDate as default};