+++ lib/number/round.ts
@@ -8,7 +8,7 @@ const ROUND_EPSILON = 1 + Number.EPSILON;
  */
 function round (val:number, precision:number = 0):number {
     if (!Number.isFinite(val)) throw new TypeError('Value should be numeric');
-    if (!Number.isInteger(precision) || precision <= 0) return Math.round(val * ROUND_EPSILON);
+    if (typeof precision !== 'number' || precision <= 0 || (precision | 0) !== precision) return Math.round(val * ROUND_EPSILON);
 
     const exp = Math.pow(10,  precision);
     return Math.round((val * exp) * ROUND_EPSILON)/exp;